+++ src/base/static/client/mapseed-api-client/activity.js
