+++ src/base/static/client/mapseed-api-client/attachments.js
